+++ deps.ts
@@ -16,7 +16,4 @@ export {
 
 export { createHelpMenu } from "https://raw.githubusercontent.com/drashland/services/master/cli/help_menu_service.ts";
 
-export {
-  commandIs,
-  commandRequiresArgs,
-} from "https://raw.githubusercontent.com/drashland/services/master/cli/command_service.ts";
+export { CommandService } from "https://raw.githubusercontent.com/drashland/services/master/cli/command_service.ts";
